error catching.

+++ panel.manageDataset.php
@@ -113,23 +113,32 @@
 				$colorString2 = 'null';
 			}
 			$parentFile   = "users/".$user."/projects/".$project."/parent.txt";
-			$handle       = fopen($parentFile,'r');
-			$parentString = trim(fgets($handle));
-			fclose($handle);
-			// getting project name
-			$projectNameString = file_get_contents("users/".$user."/projects/".$project."/name.txt");
-			$projectNameString = trim($projectNameString);
+			if (file_exists($parentFile)) {
+				$handle       = fopen($parentFile,'r');
+				$parentString = trim(fgets($handle));
+				fclose($handle);
+				// getting project name
+				$projectNameString = file_get_contents("users/".$user."/projects/".$project."/name.txt");
+				$projectNameString = trim($projectNameString);
 
-			$key = $key_;
-			echo "\n\t\t\t\t<!-- project '{$project}', #{$key}. --!>\n\t\t\t\t";
-			echo "<span id='p_label_".$key."' style='color:#CC0000;'>\n\t\t\t\t";
-			echo "<font size='2'>".($key+1).".";
-			echo "<button id='project_delete_".$key."' type='button' onclick=\"parent.deleteProjectConfirmation('".$user."','".$project."','".$key."')\">Delete</button>";
+				$key = $key_;
+				echo "\n\t\t\t\t<!-- project '{$project}', #{$key}. --!>\n\t\t\t\t";
+				echo "<span id='p_label_".$key."' style='color:#CC0000;'>\n\t\t\t\t";
+				echo "<font size='2'>".($key+1).".";
+				echo "<button id='project_delete_".$key."' type='button' onclick=\"parent.deleteProjectConfirmation('".$user."','".$project."','".$key."')\">Delete</button>";
 
-			echo $projectNameString;
-			echo "</font></span>\n\t\t\t\t";
-			echo "<span id='p_delete_".$key."'></span><br>\n\t\t\t\t";
-			echo "<div id='frameContainer.p3_".$key."'></div>\n";
+				echo $projectNameString;
+				echo "</font></span>\n\t\t\t\t";
+				echo "<span id='p_delete_".$key."'></span><br>\n\t\t\t\t";
+				echo "<div id='frameContainer.p3_".$key."'></div>\n";
+			} else {
+				// an error has happened.
+				$key = $key_;
+				echo "\n\t\t\t\t<!-- project '{$project}', #{$key}. --!>\n\t\t\t\t";
+				echo "<span id='p_label_".$key."' style='color:#333333;'>\n\t\t\t\t";
+				echo "<font size='2'>".($key+1).".";
+				echo "<button id='project_delete_".$key."' type='button' onclick=\"parent.deleteProjectConfirmation('".$user."','".$project."','".$key."')\">Delete</button>";
+			}
 		}
 		foreach($projectFolders_working as $key_=>$project) {
 			// Load colors for project.
@@ -143,24 +152,28 @@
 				$colorString1 = 'null';
 				$colorString2 = 'null';
 			}
-			$parentFile   = "users/".$user."/projects/".$project."/parent.txt";
-			$handle       = fopen($parentFile,'r');
-			$parentString = trim(fgets($handle));
-			fclose($handle);
-			// getting project name
-			$projectNameString = file_get_contents("users/".$user."/projects/".$project."/name.txt");
-			$projectNameString = trim($projectNameString);
+			if (file_exists($parentFile)) {
+				$parentFile   = "users/".$user."/projects/".$project."/parent.txt";
+				$handle       = fopen($parentFile,'r');
+				$parentString = trim(fgets($handle));
+				fclose($handle);
+				// getting project name
+				$projectNameString = file_get_contents("users/".$user."/projects/".$project."/name.txt");
+				$projectNameString = trim($projectNameString);
 
-			$key = $key_ + $userProjectCount_starting;
-			echo "\n\t\t\t\t<!-- project '{$project}', #{$key}. --!>\n\t\t\t\t";
-			echo "<span id='p_label_".$key."' style='color:#BB9900;'>\n\t\t\t\t";
-			echo "<font size='2'>".($key+1).".";
-			echo "<button id='project_delete_".$key."' type='button' onclick=\"parent.deleteProjectConfirmation('".$user."','".$project."','".$key."')\">Delete</button>";
-			echo $projectNameString;
+				$key = $key_ + $userProjectCount_starting;
+				echo "\n\t\t\t\t<!-- project '{$project}', #{$key}. --!>\n\t\t\t\t";
+				echo "<span id='p_label_".$key."' style='color:#BB9900;'>\n\t\t\t\t";
+				echo "<font size='2'>".($key+1).".";
+				echo "<button id='project_delete_".$key."' type='button' onclick=\"parent.deleteProjectConfirmation('".$user."','".$project."','".$key."')\">Delete</button>";
+				echo $projectNameString;
 
-			echo "</font></span>\n\t\t\t\t";
-			echo "<span id='p_delete_".$key."'></span><br>\n\t\t\t\t";
-			echo "<div id='frameContainer.p2_".$key."'></div>\n";
+				echo "</font></span>\n\t\t\t\t";
+				echo "<span id='p_delete_".$key."'></span><br>\n\t\t\t\t";
+				echo "<div id='frameContainer.p2_".$key."'></div>\n";
+			} else {
+				// an error has happened.
+			}
 		}
 ?>
 <script type='text/javascript'>
